// src/app/interfaces/saude.interface.ts
export interface Saude{
  key$?:string;
  // nome: string;
  // sexo: string;
  // infantil: string;
  // naturalidade: string;
  // uf: string;
  // dataNascimento: string;
  // identidade: string;
  // endereco: string;
  // num: string;
  // complemento: string;
  // bairro: string;
  // cep: string;
  // telefone: string;
  problemaGraveSaude: string;
  remedioControlado: string;
  planoSaude: string;
  alergia: string;
  remedioHabitual: string;
  acidente: string;
  contato1nome: string;
  contato1tel: string;
  contato2nome: string;
  contato2tel: string;
}